+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0.4] - 2025-06-20
+
 ### Added
 
 - Release notes as output in the GitHub Action (#38)
@@ -42,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial project skeleton (#1)
 - Initial Action Code
 
-[unreleased]: https://github.com/baynezy/ChangeLogger.Action/compare/1.0.0.3...HEAD
+[unreleased]: https://github.com/baynezy/ChangeLogger.Action/compare/1.1.0.4...HEAD
+[1.1.0.4]: https://github.com/baynezy/ChangeLogger.Action/compare/1.0.0.3...1.1.0.4
 [1.0.0.3]: https://github.com/baynezy/ChangeLogger.Action/compare/0.0.6.10...1.0.0.3
 [0.0.6.10]: https://github.com/baynezy/ChangeLogger.Action/compare/0.0.5.9...0.0.6.10
 [0.0.5.9]: https://github.com/baynezy/ChangeLogger.Action/compare/0.0.4.8...0.0.5.9